+++ src/game/Tile.java
@@ -4,15 +4,15 @@ package game;
 // 3:local visited, 4; global visited
 
 public class Tile {
-    public final int q;
-    public final int r;
-    public int state;
+    public final byte q;
+    public final byte r;
+    public byte state;
     public Boolean playable;
 
     public Tile(int q, int r, int state, boolean playable){
-        this.q  = q;
-        this.r = r;
-        this.state = state;
+        this.q  = (byte)q;
+        this.r = (byte)r;
+        this.state = (byte)state;
         this.playable = playable;
     }
 }
